Move the implementation logic into the pipes Fix a typo in the image documentation
